be more careful about keeping nameSource accurate
we already updated parentOption.nameSource when clicking up/down on
the name source option, but that's not the only way to change the index
of the name source. notably, moving an adjacent child option up or down
can do can also move the name source, effectively usurping it. if the "new"
name source isn't actually a valid name source (e.g. it's a toggle option),
then WeakAuras correctly barfs on the garbage data.

Fixes #5074