Change Profiling frame resize/toggle behaviour
This change aims to enhance the behavior of the profiling frame with the following updates:

When right-clicking the minimap icon, if the window is shown, profiling will stop (if running) and then hide. This aligns the action with the tooltip description.

The resize button will now have bounds based on MinPanelWidth and MinPanelHeight. Resizing the window excessively previously caused the buttons and scroll frame to exceed their parent’s boundaries.

The minimize button will now reduce the frame to a smaller, more practical window, as expected, according to the variables MinPanelMinimizedHeight and MinPanelMinimizedWidth.

The maximize button will ensure the window is maximized to either its previous height (if greater than MinPanelWidth) or to MinPanelWidth. (this may be redundant due to the resize bounds)