TWW+Model SubElement: Replace "Clips by Progress" by "Anchor to FG"
In TWW, the SetClipsChildren api no longer clips PlayerModels, thus the
old implementation for making the model be on the foreground no longer
works.

Replace that option with the similar directly attaching of the model to
the foreground. This does look worse but after tinkering for a good bit,
that seems to be the only possibility.