be more careful about adding new effective spell ids the watched field contains a list of spell ids, and is consumed by SendEventsForSpell, invokes ScanEventsByID for each item in the array. But we add an item to the array every time a new effective spell id is detected, so this array can grows very large & contain many duplicates if a spell is frequently replaced, causing a cooldown triggerFunc that tracks such a spell to be run many many times for a single SPELL_UPDATE_COOLDOWN event
So, check first if the array has the spell id before adding it. This still seems slightly wrong, but naievely swapping the array for a table<number, true> introduced
problems with scheduled checks that proved elusive to fix for me as a relative outsider to this bit of code.

(partially) Fixes #5380