Fix error message "Attempt to call table value"
This was a regression introduced by confussing variable names. Usually
the return value of LoadFunction is a function, except for custom
variables where the loadedFunction is instead a table value.

Furthermore the only place where we validate is for custom variables, so
we just need to pass the loadedFunction onwards to the validator.