Print a warning on finding empty settings
If the game corrupts the saved settings, we automatically take a
snapshot because db.dbVersion is nil.

This has lead to users being confussed, whose settings were corrupted,
because they wrongly assume that the snapshotting deleted the auras.

Instead, if we find entirely empty settings print a message pointing to
the discord.