Classic Era: Fix Minimize Button
Somehow the TitleContainer ended up in front of the Minimize Container
on classic era (but not on any other variant), and only in the latest
patch.

I have no idea what actually changed, but puting the minimize button
before the title container is probably a good idea on all versions.

Fixes: #5647