Fix font justify missing after update
If an update changes both the position and the justify setting, then
somehow the justify setting did not get applied.

This is almost certainly a bug in how wow optimizes the text layout,
apparently the text layout is not properly marked as dirty.

Settting the text to an empty string, which is later overriden with the
correct text seems to be enough to force wow to relayout the text

Fixes: #5654