Tweak cloneId matching between triggers
When multiple triggers provide state then their states (or more
accurately their cloneIds) are matched.

Tweak the rules a bit to benefit BT2. If an exact match to a cloneId is
not found in the other triggers, then the "prefix" of the cloneId is
tried next. The prefix is the string before the first "."

In "Combine matches per unit" the cloneId remains the unit.

If the "Combine matches per unit" is not set, BT2 sets the cloneId's
prefix to the unit.

Thus if the BT2 trigger is the dynamic information providing trigger,
and other triggers use unit based cloneIds those get matched.