Make state.show optional
Change how custom code (and internal code) indicates that a state should
be shown. Instead of relying on the magic property "show", the
existenance of a state is enough.

Adding a state to the allstates table shows it, removing a state from
allstates hides it.

Thus setting state.show to true or false is no longer necessary,
although for backwards compability setting show to false still hides
clones.

Setting show to nil was never intended but had the same effect as
setting it to false. For existing aura a compability flag is set,
although if a custom TSU updater sets show to nil, the aura gets a warning.

This also fixes #5942, because the states are now removed from allstates
before the progress update is run. Thus the stale information is no
longer used.

Fixes: #5942