Fix some problems that appear on the Titan Forged server.
Completed missing Chinese translations.
The Glyph options should not be shown on clients other than Mists.
Player characters on Vanilla, BCC, and Titan Forged servers will not be in the Neutral faction.
The first raid tier on Titan Forged Phase 1 is Molten Core.
Some API functions or variable names are not applicable on Titan Forged servers.